+++ packages/extractor/src/extract/meta/ffprobe.js
@@ -7,19 +7,8 @@ const { toPipe, conditionalTask } = require('../../stream/task');
 
 const ffprobeSuffix = 'ffprobe.json';
 
-const getFfprobePath = useNative => {
-  if (useNative.includes('ffprobe')) {
-    log.debug('Use native system command ffprobe')
-    return getNativeCommand('ffprobe')
-  } else {
-    return require('@ffprobe-installer/ffprobe').path
-  }
-}
-
-function videoMeta(storage, options) {
+function videoMeta(storage, ffprobePath) {
   const test = entry => entry.type === 'video' && !storage.hasEntryFile(entry, ffprobeSuffix);
-
-  const ffprobePath = getFfprobePath(options.useNative)
 
   const task = (entry, cb) => {
     const t0 = Date.now();